refactor(core): hoist visible-index nav helpers from pptx to share with xlsx
Move the format-agnostic nextVisibleIndex/resolveVisibleIndex (pure index math
over an isHidden callback) from packages/pptx into @silurus/ooxml-core so the
xlsx hidden-sheet skip navigation reuses them instead of duplicating. pptx
re-exports them from core; selectHidden (Slide-specific) stays in pptx.
